Gaps in game field fix

+++ app/src/main/java/ru/reosfire/minesweeper/field/cells/Cell.kt
@@ -3,5 +3,5 @@ package ru.reosfire.minesweeper.field.cells
 import android.graphics.Canvas
 
 abstract class Cell {
-    abstract fun renderTo(canvas: Canvas, xStart: Float, yStart: Float, xEnd: Float, yEnd: Float)
+    abstract fun renderTo(canvas: Canvas, xStart: Int, yStart: Int, xEnd: Int, yEnd: Int)
 }
